fix warnings

+++ MarioBros3/TileMap.cpp
@@ -27,9 +27,9 @@ void TileMap::LoadMapResource()
 	LPTEXTURE texMap = texture->Get(id);
 
 	int tile_cell_id = FIRST_TILE_CELL_ID;
-	for (UINT i = 0; i < num_of_tileset_rows; i++)
+	for (int i = 0; i < num_of_tileset_rows; i++)
 	{
-		for (UINT j = 0; j < num_of_tileset_cols; j++)
+		for (int j = 0; j < num_of_tileset_cols; j++)
 		{
 			int spriteId = id + tile_cell_id;
 			sprites->Add(spriteId, TILE_CELL_WIDTH * j, TILE_CELL_HEIGHT * i, TILE_CELL_WIDTH * (j + 1), TILE_CELL_HEIGHT * (i + 1), texMap);
@@ -73,12 +73,12 @@ void TileMap::Render()
 	int top = (int)cy / TILE_CELL_HEIGHT;
 	int bottom = top + (GAME_SCREEN_HEIGHT / TILE_CELL_HEIGHT);
 
-	for (UINT i = top; i <=bottom ; i++)
+	for (int i = top; i <= bottom ; i++)
 	{
-		for (UINT j = left; j <= right; j++)
+		for (int j = left; j <= right; j++)
 		{
-			float x = TILE_CELL_WIDTH * j + TILE_CELL_WIDTH / 2;
-			float y = TILE_CELL_HEIGHT * i + TILE_CELL_HEIGHT / 2;
+			float x = (float)(TILE_CELL_WIDTH * j + TILE_CELL_WIDTH / 2);
+			float y = (float)(TILE_CELL_HEIGHT * i + TILE_CELL_HEIGHT / 2);
 			sprites->Get(tileMap[i][j] + id)->Draw(x, y);
 		}
 	}
